Перенесены методы из ContactCreationTests в TestBase, ApplicationManager, создан ContactHelper
