+++ addressbook-web-tests/src/test/java/ru/stqa/ptf/addressbook/appmanager/ContactHelper.java
@@ -1,0 +1,106 @@
+package ru.stqa.ptf.addressbook.appmanager;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.support.ui.Select;
+import ru.stqa.ptf.addressbook.model.ContactData;
+
+public class ContactHelper {
+    private WebDriver wd;
+
+    public ContactHelper(WebDriver wd) {
+        this.wd = wd;
+    }
+
+    public void returnToContactPage() {
+        wd.findElement(By.linkText("home page")).click();
+    }
+
+    public void initNewContact() {
+        wd.findElement(By.linkText("add new")).click();
+    }
+
+    public void fillContactData(ContactData contactData) {
+        wd.findElement(By.name("firstname")).click();
+        wd.findElement(By.name("firstname")).clear();
+        wd.findElement(By.name("firstname")).sendKeys(contactData.getFirstName());
+        wd.findElement(By.name("middlename")).click();
+        wd.findElement(By.name("middlename")).clear();
+        wd.findElement(By.name("middlename")).sendKeys(contactData.getMiddleName());
+        wd.findElement(By.name("lastname")).click();
+        wd.findElement(By.name("lastname")).clear();
+        wd.findElement(By.name("lastname")).sendKeys(contactData.getLastName());
+        wd.findElement(By.name("nickname")).click();
+        wd.findElement(By.name("nickname")).clear();
+        wd.findElement(By.name("nickname")).sendKeys(contactData.getNickName());
+        wd.findElement(By.name("title")).click();
+        wd.findElement(By.name("title")).clear();
+        wd.findElement(By.name("title")).sendKeys(contactData.getTittle());
+        wd.findElement(By.name("company")).click();
+        wd.findElement(By.name("company")).clear();
+        wd.findElement(By.name("company")).sendKeys(contactData.getCompany());
+        wd.findElement(By.name("address")).click();
+        wd.findElement(By.name("address")).clear();
+        wd.findElement(By.name("address")).sendKeys(contactData.getAddress());
+        wd.findElement(By.name("home")).click();
+        wd.findElement(By.name("home")).clear();
+        wd.findElement(By.name("home")).sendKeys(contactData.getHomePhone());
+        wd.findElement(By.name("mobile")).click();
+        wd.findElement(By.name("mobile")).clear();
+        wd.findElement(By.name("mobile")).sendKeys(contactData.getMobilePhone());
+        wd.findElement(By.name("work")).click();
+        wd.findElement(By.name("work")).clear();
+        wd.findElement(By.name("work")).sendKeys(contactData.getWorkPhone());
+        wd.findElement(By.name("fax")).click();
+        wd.findElement(By.name("fax")).clear();
+        wd.findElement(By.name("fax")).sendKeys(contactData.getFax());
+        wd.findElement(By.name("email")).click();
+        wd.findElement(By.name("email")).clear();
+        wd.findElement(By.name("email")).sendKeys(contactData.getEmail());
+        wd.findElement(By.name("email2")).click();
+        wd.findElement(By.name("email2")).clear();
+        wd.findElement(By.name("email2")).sendKeys(contactData.getEmail2());
+        wd.findElement(By.name("email3")).click();
+        wd.findElement(By.name("email3")).clear();
+        wd.findElement(By.name("email3")).sendKeys(contactData.getEmail3());
+        wd.findElement(By.name("homepage")).click();
+        wd.findElement(By.name("homepage")).clear();
+        wd.findElement(By.name("homepage")).sendKeys(contactData.getHomePage());
+        wd.findElement(By.name("bday")).click();
+        new Select(wd.findElement(By.name("bday"))).selectByVisibleText(contactData.getbDDay());
+        wd.findElement(By.xpath("//option[@value='7']")).click();
+        wd.findElement(By.name("bmonth")).click();
+        new Select(wd.findElement(By.name("bmonth"))).selectByVisibleText(contactData.getbDMonht());
+        wd.findElement(By.xpath("//option[@value='June']")).click();
+        wd.findElement(By.name("byear")).click();
+        wd.findElement(By.name("byear")).clear();
+        wd.findElement(By.name("byear")).sendKeys(contactData.getbDYear());
+        wd.findElement(By.name("aday")).click();
+        new Select(wd.findElement(By.name("aday"))).selectByVisibleText(contactData.getAnnyversaryDay());
+        wd.findElement(By.xpath("(//option[@value='7'])[2]")).click();
+        wd.findElement(By.name("amonth")).click();
+        new Select(wd.findElement(By.name("amonth"))).selectByVisibleText(contactData.getAnnyversaryMonth());
+        wd.findElement(By.xpath("(//option[@value='June'])[2]")).click();
+        wd.findElement(By.name("ayear")).click();
+        wd.findElement(By.name("ayear")).clear();
+        wd.findElement(By.name("ayear")).sendKeys(contactData.getAnnyversaryYear());
+        wd.findElement(By.name("new_group")).click();
+        new Select(wd.findElement(By.name("new_group"))).selectByVisibleText(contactData.getContactGroup());
+        wd.findElement(By.xpath("(//option[@value='7'])[2]")).click();
+        wd.findElement(By.name("address2")).click();
+        wd.findElement(By.name("address2")).clear();
+        wd.findElement(By.name("address2")).sendKeys(contactData.getAddress2());
+        wd.findElement(By.name("phone2")).click();
+        wd.findElement(By.name("phone2")).clear();
+        wd.findElement(By.name("phone2")).sendKeys(contactData.getPhone2());
+        wd.findElement(By.name("notes")).click();
+        wd.findElement(By.name("notes")).clear();
+        wd.findElement(By.name("notes")).sendKeys(contactData.getNotes());
+    }
+
+
+    public void submitContactCreation() {
+        wd.findElement(By.xpath("(//input[@name='submit'])[2]")).click();
+    }
+
+}
