+++ addressbook-web-tests/src/test/java/ru/stqa/ptf/addressbook/appmanager/GroupHelper.java
@@ -7,7 +7,6 @@ import ru.stqa.ptf.addressbook.model.GroupData;
 public class GroupHelper {
     private WebDriver wd;
 
-
     public GroupHelper(WebDriver wd) {
         this.wd = wd;
     }
