+++ addressbook-web-tests/src/test/java/ru/stqa/ptf/addressbook/tests/TestBase.java
@@ -1,6 +1,9 @@
 package ru.stqa.ptf.addressbook.tests;
 
 import org.openqa.selenium.By;
+import org.openqa.selenium.NoAlertPresentException;
+import org.openqa.selenium.NoSuchElementException;
+import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.ui.Select;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
@@ -10,6 +13,7 @@ import ru.stqa.ptf.addressbook.model.ContactData;
 public class TestBase {
 
     protected final ApplicationManager app = new ApplicationManager();
+    protected WebDriver wd;
 
     @BeforeMethod(alwaysRun = true)
     public void setUp() throws Exception {
@@ -21,4 +25,21 @@ public class TestBase {
         app.stop();
     }
 
+    private boolean isElementPresent(By by) {
+      try {
+        wd.findElement(by);
+        return true;
+      } catch (NoSuchElementException e) {
+        return false;
+      }
+    }
+
+    private boolean isAlertPresent() {
+      try {
+        wd.switchTo().alert();
+        return true;
+      } catch (NoAlertPresentException e) {
+        return false;
+      }
+    }
 }
